+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - [11428](https://github.com/vegaprotocol/vega/issues/11428) - Add buy back and treasury fee and separate discount/reward factors. 
 - [11468](https://github.com/vegaprotocol/vega/issues/11468) - Added support for volume rebate program.
+- [11523](https://github.com/vegaprotocol/vega/issues/11523) - Change method of caching to improve `AMM` snapshot performance.
 - [11459](https://github.com/vegaprotocol/vega/issues/11459) - Deprecate time weight position reward metric and replace it with time weighted notional. 
 
 ### 🐛 Fixes